# src/openapi.yaml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
  description: APIs for user management and script emotion analysis
servers:
  - url: http://localhost:3000
    description: Local server
tags:
  - name: Authentication
    description: APIs for user login and registration
  - name: ScriptAI
    description: APIs for script emotion analysis and management
paths:
  /auth/login:
    post:
      summary: Authenticate user with ID token
      tags:
        - Authentication
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: Firebase ID token
      responses:
        '200':
          description: User authenticated successfully
        '401':
          description: Invalid or expired token
        '500':
          description: Internal server error
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  description: User's first name
                lastName:
                  type: string
                  description: User's last name
                email:
                  type: string
                  description: User's email address
                password:
                  type: string
                  description: User's password
                phoneNumber:
                  type: string
                  description: User's phone number (optional)
                displayName:
                  type: string
                  description: User's display name (optional)
                photoURL:
                  type: string
                  description: URL of the user's profile picture (optional)
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Missing required fields
        '500':
          description: Internal server error
  /scripts/emotion-analysis:
    post:
      summary: Process a script for emotion analysis using Vertex AI
      tags:
        - ScriptAI
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  description: Type of script (e.g., AD)
                  example: "AD"
                region:
                  type: string
                  description: Primary country or region
                  example: "India"
                identity:
                  type: string
                  description: What the script is about (e.g., Product, Marketing, Company)
                  example: "Product"
                brand:
                  type: string
                  description: Brand name (optional)
                  example: "Lacto Calamine"
                script:
                  type: string
                  description: The content of the script (20 to 2000 words)
                  example: "Film opens with Sindhu entering a jogging track in a stadium..."
      responses:
        '200':
          description: Script processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Script processed successfully."
                  data:
                    type: object
                    properties:
                      lines:
                        type: array
                        description: Array of analyzed lines from the script
                        items:
                          type: object
                          properties:
                            lineNumber:
                              type: integer
                              example: 1
                            content:
                              type: string
                              example: "The scene opens with a sunset over a quiet beach."
                            language:
                              type: string
                              example: "English"
                            sentimentEmotion:
                              type: string
                              example: "Joy"
                            valence:
                              type: number
                              example: 9.46
                            arousal:
                              type: number
                              example: 1.3
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid input. Ensure all required fields are provided and script length is between 20 and 2000 words."
        '401':
          description: Unauthorized - No or invalid token provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No token provided."
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error."
    get:
      summary: Retrieve scripts and associated data for the authenticated user
      tags:
        - ScriptAI
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Scripts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Scripts retrieved successfully."
                  scripts:
                    type: array
                    description: Array of scripts created by the authenticated user
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: Document ID in Firestore
                          example: "documentId123"
                        userId:
                          type: string
                          description: UID of the user who created the script
                          example: "userUid123"
                        type:
                          type: string
                          description: Type of script (e.g., AD)
                          example: "AD"
                        region:
                          type: string
                          description: Primary country or region
                          example: "India"
                        identity:
                          type: string
                          description: What the script is about (e.g., Product, Marketing, Company)
                          example: "Product"
                        brand:
                          type: string
                          description: Brand name (optional)
                          example: "Lacto Calamine"
                        script:
                          type: string
                          description: Original script content
                          example: "Film opens with Sindhu entering a jogging track in a stadium..."
                        status:
                          type: string
                          description: Status of the script (e.g., pending, processed)
                          example: "processed"
                        vertexResponse:
                          type: object
                          description: Analysis results returned by Vertex AI
                          properties:
                            lines:
                              type: array
                              items:
                                type: object
                                properties:
                                  lineNumber:
                                    type: integer
                                    example: 1
                                  content:
                                    type: string
                                    example: "The scene opens with a sunset over a quiet beach."
                                  language:
                                    type: string
                                    example: "English"
                                  sentimentEmotion:
                                    type: string
                                    example: "Joy"
                                  valence:
                                    type: number
                                    example: 9.46
                                  arousal:
                                    type: number
                                    example: 1.3
                        createdAt:
                          type: string
                          format: date-time
                          description: Timestamp when the script was created
                          example: "2024-12-01T12:00:00.000Z"
        '404':
          description: No scripts found for the user
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "No scripts found for the user."
        '401':
          description: Unauthorized - No or invalid token provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No token provided."
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error."
  /auth/getToken:
    post:
      summary: Retrieve an authentication token
      description: Endpoint to retrieve an ID token using email and password from Firebase Authentication.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address.
                  example: vivek@fram3.ai
                password:
                  type: string
                  format: password
                  description: User's password.
                  example: YourPassword123
              required:
                - email
                - password
      responses:
        '200':
          description: Token retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Token retrieved successfully!
                  displayName:
                    type: string
                    description: User's Display Name
                    example: Vivek Ramchander
                  token:
                    type: string
                    description: Firebase authentication token.
                    example: eyJhbGciOiJSUzI1NiIsImtpZCI6ImZ...
                  refreshToken:
                    type: string
                    description: Firebase authentication token.
                    example: eyJhbGciOiJSUzI1NiIsImtpZCI6ImZ...
                  expiresIn:
                   type: string
                   description: Validity of token in seconds from creation
                   example: 3600 - 1Hour
        '400':
          description: Missing or invalid parameters.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Email and password are required.
        '500':
          description: Server error or Firebase authentication failure.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Failed to retrieve token.
  /auth/refreshToken:
    post:
      summary: Refresh authentication token
      description: Renew an ID token using a valid refresh token from Firebase Authentication.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: A valid refresh token issued by Firebase.
                  example: "AE4bGJv...cih8YQ"
      responses:
        '200':
          description: Token refreshed successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Token refreshed successfully!
                  token:
                    type: string
                    description: The new ID token.
                    example: "eyJhbGciOiJSUzI1NiIsImtpZCI6ImZ..."
                  refreshToken:
                    type: string
                    description: The new refresh token.
                    example: "AE4bGJv...cih8YQ"
                  expiresIN:
                    type: string
                    description: Time in seconds until the token expires.
                    example: "3600"
        '400':
          description: Missing or invalid refresh token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Refresh token is required.
        '500':
          description: Server error or failure to refresh the token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Failed to refresh token.
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT